pyproject.toml
[project]
name = "hull-tactical"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "mdomarsaleem1",email = "omarsaleem003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "numpy>=1.26,<2.0",
    "pandas (>=2.3.3,<3.0.0)",
    "transformers>=4.41,<5",
  "accelerate>=0.34,<2",
  "einops>=0.7.0,<1",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "xgboost (==2.0.3)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "prophet (==1.1.5)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "plotly (>=6.5.0,<7.0.0)",
    "chronos-forecasting",
    "torch (>=2.9.1,<3.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "cmdstanpy (>=1.3.0,<2.0.0)",
    "pystan (>=3.10.0,<4.0.0)",
    "kedro (>=1.1.1,<2.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "optuna (>=4.6.0,<5.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true
[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]
